Use build constraints
